chore(e2e): several e2e test improvements and corrections
