+++ packages/kuma-gui/features/mesh/MeshTrust.feature
@@ -4,6 +4,7 @@ Feature: mesh / mesh-identity
     Given the CSS selectors
       | Alias              | Selector                            |
       | meshtrusts-listing | [data-testid="mesh-trusts-listing"] |
+      | item             | $meshtrusts-listing tbody tr                                  |
       | summary            | [data-testid="slideout-container"]  |
 
   Scenario: MeshTrusts are listed in mesh overview
@@ -18,8 +19,8 @@ Feature: mesh / mesh-identity
       """
     When I visit the "/meshes/default" URL
     Then the "$meshtrusts-listing" element exists
-    And the "$meshtrusts-listing" element contains "trust-1"
-    And the "$meshtrusts-listing" element contains "kri_mid_default_default_foo_bar_baz"
+    And the "$item:nth-child(1)" element contains "trust-1"
+    And the "$item:nth-child(1)" element contains "kri_mid_default_default_foo_bar_baz"
 
   Scenario: Clicking on mesh trust item opens summary view
     Given the URL "/meshes/default/meshtrusts" responds with
@@ -33,7 +34,7 @@ Feature: mesh / mesh-identity
                 kri: kri_mid_default_default_foo_bar_baz
       """
     When I visit the "/meshes/default" URL
-    Then I click the "$meshtrusts-listing a:first" element
+    Then I click the "$item:nth-child(1) td:first-child a" element
     Then the URL contains "/meshes/default/overview/meshtrust/kri_mtrust_default___trust-1_"
     And the "$summary" element exists
     And the "$summary" element contains "trust-1"
@@ -59,7 +60,7 @@ Feature: mesh / mesh-identity
           - name: bar
       """
     When I visit the "/meshes/default" URL
-    Then I click the "$meshtrusts-listing tbody tr:first-child td:nth-child(3) a" element
+    Then I click the "$item:nth-child(1) td:nth-child(3) a" element
     Then the URL contains "/meshes/default/overview/meshidentity/kri_mid_default_default_foo_bar_baz"
     And the "$summary" element exists
     And the "$summary" element contains "bar"
