+++ packages/kuma-gui/features/zones/zone-cps/egresses/Item.feature
@@ -25,7 +25,7 @@ Feature: zones / egresses / item
       """
 
   Scenario: The about section has the expected content
-    When I visit the "zones/zone-cp-1/egresses/item-1/overview" URL
+    When I visit the "/zones/zone-cp-1/egresses/item-1/overview" URL
     Then the "$about-section" element exists
     And the "$about-section" element contains "kuma.io/display-name:default"
