+++ packages/kuma-gui/features/zones/zone-cps/ingresses/item/Index.feature
@@ -26,7 +26,7 @@ Feature: zones / ingresses / item
       """
 
   Scenario: The about section has the expected content
-    When I visit the "zones/zone-cp-1/ingresses/item-1/overview" URL
+    When I visit the "/zones/zone-cp-1/ingresses/item-1/overview" URL
     Then the "$about-section" element exists
     And the "$about-section" element contains "kuma.io/display-name:default"
