+++ packages/kuma-http-api/mocks/fs.ts
@@ -63,16 +63,16 @@ import _134 from './src/meshes/_/meshgateways/test-meshgateway/_'
 import _135 from './src/meshes/_/meshgateways/test-meshgateway/_rules'
 import _132 from './src/meshes/_/meshhttproutes'
 import _133 from './src/meshes/_/meshhttproutes/_'
-import _241 from './src/meshes/_/meshidentities/_'
-import _242 from './src/meshes/_/meshidentities/_/_.ts'
+import _242 from './src/meshes/_/meshidentities/_.ts'
+import _241 from './src/meshes/_/meshidentities.ts'
 import _140 from './src/meshes/_/meshmultizoneservices'
 import _141 from './src/meshes/_/meshmultizoneservices/_'
 import _171 from './src/meshes/_/meshmultizoneservices/_/_hostnames'
 import _136 from './src/meshes/_/meshservices'
 import _137 from './src/meshes/_/meshservices/_'
 import _170 from './src/meshes/_/meshservices/_/_hostnames'
-import _243 from './src/meshes/_/meshtrusts/_'
-import _244 from './src/meshes/_/meshtrusts/_/_.ts'
+import _243 from './src/meshes/_/meshtrusts'
+import _244 from './src/meshes/_/meshtrusts/_.ts'
 import _35 from './src/meshes/_/proxytemplates'
 import _36 from './src/meshes/_/proxytemplates/_'
 import _37 from './src/meshes/_/rate-limits'
