+++ packages/kuma-http-api/mocks/src/meshes/_/meshidentities/_.ts
@@ -2,92 +2,101 @@ import type { Dependencies, ResponseHandler } from '#mocks'
 import type { components } from '@kumahq/kuma-http-api'
 
 
-export default ({ fake, env }: Dependencies): ResponseHandler => (req) => {
-  const params = req.params
-  const mesh = params.mesh as string
-  const itemsCount = parseInt(env('KUMA_MESHIDENTITY_COUNT', `${fake.number.int({ min: 1, max: 3 })}`))
+export default ({ fake }: Dependencies): ResponseHandler => (req) => {
+  const kri = req.params.kri as string | undefined
+  const [
+    mesh = req.params.mesh as string,
+    _zone,
+    _namespace,
+    name = req.params.name as string,
+  ] = kri?.split('_') ?? ''
+  const k8s = req.url.searchParams.get('format') === 'kubernetes'
   return {
     headers: {},
     body: {
-      items: Array.from({ length: itemsCount }, () => ({
-        type: 'MeshIdentity',
-        mesh,
-        name: fake.word.noun(),
-        labels: {
-          'kuma.io/mesh': mesh,
-          'kuma.io/zone': fake.word.noun(),
-          'kuma.io/origin': 'zone',
-          'kuma.io/namespace': fake.word.noun(),
-        },
+      ...(k8s && { apiVersion: 'kuma.io/v1alpha1' }),
+      ...(k8s ? { kind: 'MeshIdentity' } : { type: 'MeshIdentity' }),
+      ...((() => {
+        const metadata = {
+          name,
+          labels: {
+            'kuma.io/mesh': mesh,
+            'kuma.io/zone': fake.word.noun(),
+            'kuma.io/origin': 'zone',
+            'kuma.io/namespace': fake.word.noun(),
+          }}
+        return k8s ? { metadata } : metadata
+      })()),
+      ...(!k8s && {
         creationTime: fake.date.past().toISOString(),
         modificationTime: fake.date.recent().toISOString(),
-        spec: {
-          provider: {
-            bundled: {
-              autogenerate: {
-                enabled: fake.datatype.boolean(),
-              },
-              ca: {
-                certificate: {
-                  envVar: {
-                    name: fake.word.noun() + '-cert',
-                  },
-                  file: {
-                    path: `${fake.system.directoryPath()}/${fake.system.commonFileName('crt')}`,
-                  },
-                  insecureInline: {
-                    value: fake.word.noun(),
-                  },
-                  secretRef: {
-                    kind: 'Secret',
-                    name: fake.word.noun(),
-                  },
-                  type: fake.helpers.arrayElement(['File', 'Secret', 'EnvVar', 'InsecureInline']),
-                },
-                privateKey: {
-                  envVar: {
-                    name: fake.word.noun(),
-                  },
-                  file: {
-                    path: `${fake.system.directoryPath()}/${fake.system.commonFileName('key')}`,
-                  },
-                  insecureInline: {
-                    value: fake.word.noun(),
-                  },
-                  secretRef: {
-                    kind: 'Secret',
-                    name: fake.word.noun(),
-                  },
-                  type: fake.helpers.arrayElement(['File', 'Secret', 'EnvVar', 'InsecureInline']),
-                },
-              },
-              certificateParameters: {
-                expiry: fake.date.soon().toISOString(),
-              },
-              insecureAllowSelfSigned: fake.datatype.boolean(),
-              meshTrustCreation: fake.helpers.arrayElement(['Enabled', 'Disabled']),
+      }),
+      spec: {
+        provider: {
+          bundled: {
+            autogenerate: {
+              enabled: fake.datatype.boolean(),
             },
-            spire: {
-              agent: {
-                timeout: '1s',
+            ca: {
+              certificate: {
+                envVar: {
+                  name: fake.word.noun() + '-cert',
+                },
+                file: {
+                  path: `${fake.system.directoryPath()}/${fake.system.commonFileName('crt')}`,
+                },
+                insecureInline: {
+                  value: fake.word.noun(),
+                },
+                secretRef: {
+                  kind: 'Secret',
+                  name: fake.word.noun(),
+                },
+                type: fake.helpers.arrayElement(['File', 'Secret', 'EnvVar', 'InsecureInline']),
+              },
+              privateKey: {
+                envVar: {
+                  name: fake.word.noun(),
+                },
+                file: {
+                  path: `${fake.system.directoryPath()}/${fake.system.commonFileName('key')}`,
+                },
+                insecureInline: {
+                  value: fake.word.noun(),
+                },
+                secretRef: {
+                  kind: 'Secret',
+                  name: fake.word.noun(),
+                },
+                type: fake.helpers.arrayElement(['File', 'Secret', 'EnvVar', 'InsecureInline']),
               },
             },
-            type: fake.helpers.arrayElement(['Bundled', 'Spire']),
+            certificateParameters: {
+              expiry: fake.date.soon().toISOString(),
+            },
+            insecureAllowSelfSigned: fake.datatype.boolean(),
+            meshTrustCreation: fake.helpers.arrayElement(['Enabled', 'Disabled']),
           },
-          selector: {
-            dataplane: {
-              matchLabels: {
-                'kuma.io/mesh': mesh,
-                'kuma.io/zone': fake.word.noun(),
-              },
+          spire: {
+            agent: {
+              timeout: '1s',
             },
           },
-          spiffeID: {
-            path: fake.kuma.spiffeId({ mesh }),
-            trustDomain: fake.word.noun(),
+          type: fake.helpers.arrayElement(['Bundled', 'Spire']),
+        },
+        selector: {
+          dataplane: {
+            matchLabels: {
+              'kuma.io/mesh': mesh,
+              'kuma.io/zone': fake.word.noun(),
+            },
           },
         },
-      }) satisfies components['schemas']['MeshIdentityItem']),
+        spiffeID: {
+          path: fake.kuma.spiffeId({ mesh }),
+          trustDomain: fake.word.noun(),
+        },
+      } satisfies components['schemas']['MeshIdentityItem']['spec'],
     },
   }
 }
