+++ packages/kuma-http-api/mocks/src/meshes/_/meshtrusts/_.ts
@@ -2,43 +2,54 @@ import type { Dependencies, ResponseHandler } from '#mocks'
 import type { components } from '@kumahq/kuma-http-api'
 
 
-export default ({ fake, env }: Dependencies): ResponseHandler => (req) => {
-  const params = req.params
-  const mesh = params.mesh as string
-  const k8s = env('KUMA_ENVIRONMENT', 'universal') === 'kubernetes'
-  const itemsCount = parseInt(env('KUMA_MESHTRUST_COUNT', `${fake.number.int({ min: 1, max: 3 })}`))
+export default ({ fake }: Dependencies): ResponseHandler => (req) => {
+  const kri = req.params.kri as string | undefined
+  const [
+    mesh = req.params.mesh as string,
+    _zone,
+    _namespace,
+    name = req.params.name as string,
+  ] = kri?.split('_') ?? ''
+  const k8s = req.url.searchParams.get('format') === 'kubernetes'
   const namespace = fake.word.noun()
   const zone = fake.word.noun()
+  const origin = fake.kuma.origin()
+
   return {
-    headers: {
-    },
+    headers: {},
     body: {
-      items: Array.from({ length: itemsCount }, () => ({
-        type: 'MeshTrust',
-        mesh,
-        name: fake.word.noun(),
-        labels: {
-          'k8s.kuma.io/namespace': fake.word.noun(),
-          'kuma.io/env': k8s ? 'kubernetes' : 'universal',
-          'kuma.io/mesh': mesh,
-          'kuma.io/origin': origin,
-          ...(origin === 'zone' && { 'kuma.io/zone': zone }),
-        },
+      ...(k8s && { apiVersion: 'kuma.io/v1alpha1' }),
+      ...(k8s ? { kind: 'MeshTrust' } : { type: 'MeshTrust' }),
+      ...((() => {
+        const metadata = {
+          mesh,
+          name,
+          labels: {
+            'k8s.kuma.io/namespace': fake.word.noun(),
+            'kuma.io/env': k8s ? 'kubernetes' : 'universal',
+            'kuma.io/mesh': mesh,
+            'kuma.io/origin': origin,
+            ...(origin === 'zone' && { 'kuma.io/zone': zone }),
+          },
+        }
+        return k8s ? { metadata } : metadata
+      })()),
+      ...(!k8s && {
         creationTime: fake.date.past().toISOString(),
         modificationTime: fake.date.recent().toISOString(),
-        spec: {
-          caBundles: [{
-            pem: {
-              value: fake.kuma.certificate(false),
-            },
-            type: 'Pem',
-          }],
-          origin: {
-            kri: fake.kuma.kri({ shortName: 'mid', mesh, namespace, zone }),
+      }),
+      spec: {
+        caBundles: [{
+          pem: {
+            value: fake.kuma.certificate(false),
           },
-          trustDomain: `${mesh}.${fake.word.noun()}.mesh.local`,
+          type: 'Pem',
+        }],
+        origin: {
+          kri: fake.kuma.kri({ shortName: 'mid', mesh, namespace, zone }),
         },
-      }) satisfies components['schemas']['MeshTrustItem']),
+        trustDomain: `${mesh}.${fake.word.noun()}.mesh.local`,
+      } satisfies components['schemas']['MeshTrustItem']['spec'],
     },
   }
 }
